code/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Cougar Tasker <cougartasker@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">= 3.10, < 3.12"
numpy = "^1.26.1"
toml = "^0.10.2"
schema = "^0.7.5"
immutables = "^0.20"
customtkinter = "^5.2.1"
pillow = "^10.1.0"
colour = "^0.1.5"
numba = "^0.58.1"
tkinter-tooltip = "^2.1.0"
packaging = "^23.2"
pyside6 = "^6.6.1"


[tool.poetry.scripts]
start = "src.main:main"
profile = "src.profile:profile"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
mypy = "^1.6.1"
pytest-mock = "^3.12.0"
types-toml = "^0.10.8.7"
pep8-naming = "^0.13.3"
wemake-python-styleguide = "^0.18.0"
snakeviz = "^2.2.0"
mkdocs-material = "^9.5.0"
pytz = "^2023.3.post1"
mkdocstrings = "^0.24.0"
mkdocs-gen-files = "^0.5.0"
mkdocs-literate-nav = "^0.6.1"
mkdocs-section-index = "^0.3.8"
mkdocstrings-python = "^1.7.5"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 80

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 80

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
